+++ opentelemetry/src/main/scala/zio/telemetry/opentelemetry/baggage/propagation/BaggagePropagator.scala
@@ -2,7 +2,7 @@ package zio.telemetry.opentelemetry.baggage.propagation
 
 import io.opentelemetry.api.baggage.propagation.W3CBaggagePropagator
 import io.opentelemetry.context.propagation.TextMapPropagator
-import zio.telemetry.opentelemetry.context.internal.Propagator
+import zio.telemetry.opentelemetry.context.Propagator
 
 trait BaggagePropagator extends Propagator
